crossplatformability improvements

+++ source/tv_server/tv_module/tvm_pmove.c
@@ -161,12 +161,12 @@ static void TVM_PM_ClampAngles( void )
 			if( temp > (short)ANGLE2SHORT( 90 ) - 1 )
 			{
 				pm->playerState->pmove.delta_angles[i] = ( ANGLE2SHORT( 90 ) - 1 ) - pm->cmd.angles[i];
-				temp = ANGLE2SHORT( 90 ) - 1;
+				temp = (short)ANGLE2SHORT( 90 ) - 1;
 			}
 			else if( temp < (short)ANGLE2SHORT( -90 ) + 1 )
 			{
 				pm->playerState->pmove.delta_angles[i] = ( ANGLE2SHORT( -90 ) + 1 ) - pm->cmd.angles[i];
-				temp = ANGLE2SHORT( -90 ) + 1;
+				temp = (short)ANGLE2SHORT( -90 ) + 1;
 			}
 		}
